Split data model structures out of datastore.rs
