+++ crates/noseyparker/src/datastore/finding_data.rs
@@ -1,0 +1,27 @@
+use crate::blob_metadata::BlobMetadata;
+use crate::match_type::Match;
+use crate::provenance_set::ProvenanceSet;
+
+use super::MatchId;
+use super::Status;
+
+// -------------------------------------------------------------------------------------------------
+// FindingData
+// -------------------------------------------------------------------------------------------------
+/// A set of match data entries
+pub type FindingData = Vec<FindingDataEntry>;
+
+// -------------------------------------------------------------------------------------------------
+// FindingDataEntry
+// -------------------------------------------------------------------------------------------------
+/// Data for a single `Match`
+#[derive(Debug)]
+pub struct FindingDataEntry {
+    pub provenance: ProvenanceSet,
+    pub blob_metadata: BlobMetadata,
+    pub match_id: MatchId,
+    pub match_val: Match,
+    pub match_comment: Option<String>,
+    pub match_score: Option<f64>,
+    pub match_status: Option<Status>,
+}
