+++ crates/noseyparker/src/datastore/finding_summary.rs
@@ -1,0 +1,25 @@
+use serde::Serialize;
+
+// -------------------------------------------------------------------------------------------------
+// FindingSummary
+// -------------------------------------------------------------------------------------------------
+
+/// A summary of matches in a `Datastore`.
+#[derive(Serialize)]
+pub struct FindingSummary(pub Vec<FindingSummaryEntry>);
+
+#[derive(Serialize)]
+pub struct FindingSummaryEntry {
+    pub rule_name: String,
+    pub distinct_count: usize,
+    pub total_count: usize,
+}
+
+impl std::fmt::Display for FindingSummary {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        for entry in self.0.iter() {
+            writeln!(f, "{}: {} ({})", entry.rule_name, entry.distinct_count, entry.total_count)?;
+        }
+        Ok(())
+    }
+}
